OJ-3267: Move date-time util
